+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Basic VDF file syntax
 - VDF objects folding
 - VDF structure view and structure-aware navigation bar
+- VDF formatter and basic code style settings